More complete Mac OS X support
This adds installation for Macs via dmg, which
are actually a pkg-inside-dmg and therefore require
mounting, then installation.

There's more to be done; specifically this -only-
installs Puppet and therefore does not leave you
with a working setup -- facter and hiera are
separate downloads and there's no automatic dependency
resolution in Apple-land.